requirements.txt
python-binance==1.0.15
numpy==1.25.0
python-dotenv==1.0.0